+++ game_project_perditio/per_world.cpp
@@ -72,7 +72,7 @@ void PERWorld::ObjectsGraphicsUpdate(PERAudio& audio, double dTime)
 
 void PERWorld::Render(PERRenderer& renderer, double frameGap)
 {
-	for (int i = 0; i < m_sortedObjects.size(); ++i) {
+	for (size_t i = 0; i < m_sortedObjects.size(); ++i) {
 		m_sortedObjects[i]->GetGraphics().Render(*m_sortedObjects[i], renderer, frameGap);
 	}
 	GetHud().RendererInWorld(renderer);
